Move Limit selector out to separate component
